feature: reorganize go-util's cli
